+++ cerberus/__init__.py
@@ -1,7 +1,12 @@
+# Setting up logging here because this is the root of the module
+import logging
+
+# This avoids the "No handler found" warnings.
+logging.getLogger(__name__).addHandler(logging.NullHandler())
+
 __all__ = ['client', 'user_auth', 'aws_auth', 'util']
 
-CLIENT_VERSION = '2.1.2'
-
+CLIENT_VERSION = '2.2.0'
 
 class CerberusClientException(Exception):
     """Wrap third-party exceptions expected by the Cerberus client."""